ORMización `Dinero`

+++ src/main/java/ar/edu/utn/frba/dds/models/repositories/recompensas/CanjeosRepository.java
@@ -26,7 +26,7 @@ public class CanjeosRepository extends HibernateEntityManager<Canjeo, Long> {
     return em.createQuery(query).getSingleResult();
   }
 
-  public double getPuntosDisponibles(Colaborador colaborador) {
+  public long getPuntosDisponibles(Colaborador colaborador) {
     long puntosTotales = CalculadoraDePuntos.calcular(colaborador);
     long puntosGastados = findPuntosGastados(colaborador);
